Single version of TSP 42 working

+++ src/main/groovy/island_model/ConvergedRecord.groovy
@@ -5,7 +5,7 @@ class ConvergedRecord {
   int findingNode
   int generationsTaken
   int instance
-  double seedValue
+  long seedValue
 
   String toString(){
     String s = "Fitness: ${convergedIndividual.getFitness()}, " +
